requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
facenet-pytorch==2.5.2
torch==2.3.1+cpu
torchvision==0.18.1+cpu
pillow==11.3.0
numpy==1.26.4
python-multipart==0.0.20
requests==2.32.4
typing_extensions==4.14.1
starlette==0.47.2
anyio==4.9.0
h11==0.16.0
sniffio==1.3.1
click==8.2.1
pydantic==2.11.7
pydantic_core==2.33.2
